+++ client/src/App.js
@@ -4,7 +4,7 @@ import LandingPage from './components/LandingPage/LandingPage'
 import Register from './components/Register/Register'
 import Login from './components/Login/Login';
 import Home from './components/Home/Home';
-import NavBar from './components/NavBar/NavBar';
+import DetailClinica from './components/DetailClinica/DetailClinica';
 import AboutUs from './components/AboutUs/AboutUs';
 
 
@@ -17,6 +17,7 @@ function App() {
           <Route exact path='/register' element={<Register/>}/>
           <Route exact path='/login' element={<Login />}/>
           <Route path='/Home' element={<Home/>}/>
+          <Route path='/home/clinica/:name' element={<DetailClinica/>}/>
           <Route path='/AboutUs' element={<AboutUs/>}/>
         </Routes>
       </div>
